// src/app/models/settings.ts
export class Settings {
  activeYear: boolean;
  firstName: string;
  lastName: string;
  birthDate: string;
  taxyear: string;
  enablebudgets: boolean;
  enableupdates: boolean;
  enableportfolios: boolean;
  enablebonds: boolean;
  enablesavings: boolean;
  enableoneoffs: boolean;
  enablepensions: boolean;
  enablesalaries: boolean;
  enableproperty: boolean;
  enableloans: boolean;
  enabledebts: boolean;
  Y1CPIRate: number;
  Y2CPIRate: number;
  Y3CPIRate: number;
  Y4CPIRate: number;
  Y5CPIRate: number;
  Y6CPIRate: number;
  Y7CPIRate: number;
  Y8CPIRate: number;
  Y9CPIRate: number;
  Y20CPIRate: number;
  Y30CPIRate: number;
  dividendRate: number;
  investmentGrowthRate: number;
  propertyGrowthRate: number;
  Y1PensionRate: number;
  Y2PensionRate: number;
  Y3PensionRate: number;
  Y4PensionRate: number;
  Y5PensionRate: number;
  Y6PensionRate: number;
  Y7PensionRate: number;
  Y8PensionRate: number;
  Y9PensionRate: number;
  Y20PensionRate: number;
  Y30PensionRate: number;
  ordercount: number;
}